null check for subnav dropdown
