+++ scripts/utils.js
@@ -3,3 +3,15 @@ export function createElemWithClass(type, ...classNames) {
   elem.classList.add(...classNames);
   return elem;
 }
+
+function setAttributes(element, attributes) {
+  Object.keys(attributes).forEach((attributeKey) => {
+    element.setAttribute(attributeKey, attributes[attributeKey]);
+  });
+}
+
+export function createElementWithAttributes(tag, attributes) {
+  const element = document.createElement(tag);
+  setAttributes(element, attributes);
+  return element;
+}
